Refactors MainMenu and adds isCollapsed to its state.
Adds top level comment to Main Menu explaining its responsibilities and states.

Adds isCollapsed state so that the Main Menu can be altered when the screen size changes. More
on this later.

Uses isCollapsed state in render to determine if a button is needed to toggle the main menu.

Refactors rendering call if mainMenuOpen is true into a MainMenuList. This new component will
receive this.state.isCollapsed as a prop and conditionally render the list that is needed for
the main menu. Refactoring also cleaned up MainMenu.

Removes unnecessary browseMenu variable and replaces its references with menu instead.