Reorganized Layout code to increase separation of concerns and testability
The Layout and navigation state managment system needed to be rewritten in order to
be tested. The data that some of the navigation menu components needed was encapsulated
in those components and couldn't be controlled. In addition, the Layout component was
doing too many things at one time and was broken up to decreatesting difficulty.

Also realized that the HeaderModalLeft state management code didn't need to be
located at the top level, because the HeaderModalLeft portal is always appended to
the body of the document.

First, to delegate the responsibility of changing the rendered menu when the screen
size changes, the PureHeader and Header components were implemented. Pure Header takes
all of the data that is needed for the header navigation and passes it to the menu
that should be rendered at the screen size.

Next, the modal management code was split between CollapsedMainMenu and ExpandedMainMenu. This
helped the components make more sense semantically, and also delegated the modal management code
down the tree, but resulted in code duplication that needs to be addresssed. A similar pattern was
followed for the CategoryMenu and NestedCategoryMenu components, which were split from the
withSubcategoryMenu HOC.

All tests were adjusted to account for the change in components and are passing. Tests
for the overall navigation scheme still need to be implemented after code duplication is addressed.