Removes code duplication from CollapsedMainMenu and ExpandedMainMenu
NestedCategoryMenu was deleted in favor of extracting the main menu button
to CollapsedMainMenu and reusing CategoryMenu instead. This caused a small
change in the UI; now the "Main Menu" button is shown in both nested menus
in CollapsedMainMenu.

Extracted HeaderModalLeft state management code from CollapsedMainMenu and
ExpandedMainMenu to withHeaderModalLeft. This new HOC manages the state and
provides reusable state changing functions for HeaderModalLeft.

Replaces references to CollapsedMainMenu and ExpandedMainMenu in Header with
calls to withHeaderModalLeft.

Gave the unordered list in HeaderModalLeft more padding because some of its
list items were being rendered out of the content div.