Implements features that were previously handled by the conversion script
Adds text formatting methods to the util folder to titleCase strings and
prettify the categories and subcategories for display in the category menu.

Prettifies the Categories and Subcategories in category-menu.js

Implements menu button sorting in CategoryMenu, so the buttons are
sorted by their prettified text, rather than their original text.

Modifies props of DDMenuBtn to accept a text prop that is used to sort,
rather than relying on the children prop. In case a button includes an
icon and some text or something.

Implements tests for newly implemented methods and modifies existing tests to
pass under new conditions.