Implements Node JSON production from TSV
Wanted JSON production to be more streamlined for Craig and myself, and needed
a good way to clean up the data, extract details from descript2, and create a set
of data for each page of items and for each item, so all of the python tsv
conversion features were scripted out in node in the data-processing folder.

Now both sets of data are produced from files in /raw/ and are output to the
src/data.