Modifies Carousel props
Old implementation of carousel was limited to a set of photos that all use the same alt text.
Also was limited to only displaying photos. Need to be able to use different alt text and
render some html over top of images in the slideshow that require it. So Carousel props
were modified to take an array of content objects rather than an alt and an array of photos.

Revises files that rely on Carousel and also updates carousel tests.